removed code from switch and gave them individual methods for easier readability

+++ LexiconExercise1/Employee.cs
@@ -9,9 +9,9 @@ namespace LexiconExercise1
     internal class Employee
     {
         public string Name { get; set; }
-        public double Salary { get; set; }
+        public decimal Salary { get; set; }
 
-        public Employee(string name, double salary)
+        public Employee(string name, decimal salary)
         {
             Name = name;
             Salary = salary;
